requirements.txt
numpy
mxnet-cu90
matplotlib
pillow